+++ tests/unit/utils.py
@@ -1,10 +1,11 @@
 import os
 import tempfile
+from copy import copy, deepcopy
 
 from hdmf.build import TypeMap
 from hdmf.container import Container
-from hdmf.spec import NamespaceCatalog, SpecCatalog, SpecNamespace, NamespaceBuilder
-from hdmf.utils import docval, getargs
+from hdmf.spec import GroupSpec, DatasetSpec, NamespaceCatalog, SpecCatalog, SpecNamespace, NamespaceBuilder
+from hdmf.utils import docval, getargs, get_docval
 
 CORE_NAMESPACE = 'test_core'
 
@@ -153,3 +154,102 @@ def create_load_namespace_yaml(namespace_name, specs, output_dir, incl_types, ty
     ns_builder.export(ns_filename, outdir=output_dir)
     ns_path = os.path.join(output_dir, ns_filename)
     type_map.load_namespaces(ns_path)
+
+
+# ##### custom spec classes #####
+
+def swap_inc_def(cls, custom_cls):
+    args = get_docval(cls.__init__)
+    ret = list()
+    for arg in args:
+        if arg['name'] == 'data_type_def':
+            ret.append({'name': 'my_data_type_def', 'type': str,
+                        'doc': 'the NWB data type this spec defines', 'default': None})
+        elif arg['name'] == 'data_type_inc':
+            ret.append({'name': 'my_data_type_inc', 'type': (custom_cls, str),
+                        'doc': 'the NWB data type this spec includes', 'default': None})
+        else:
+            ret.append(copy(arg))
+    return ret
+
+
+class BaseStorageOverride:
+    __type_key = 'my_data_type'
+    __inc_key = 'my_data_type_inc'
+    __def_key = 'my_data_type_def'
+
+    @classmethod
+    def type_key(cls):
+        ''' Get the key used to store data type on an instance'''
+        return cls.__type_key
+
+    @classmethod
+    def inc_key(cls):
+        ''' Get the key used to define a data_type include.'''
+        return cls.__inc_key
+
+    @classmethod
+    def def_key(cls):
+        ''' Get the key used to define a data_type definition.'''
+        return cls.__def_key
+
+    @classmethod
+    def build_const_args(cls, spec_dict):
+        """Extend base functionality to remap data_type_def and data_type_inc keys"""
+        spec_dict = copy(spec_dict)
+        proxy = super(BaseStorageOverride, cls)
+        if proxy.inc_key() in spec_dict:
+            spec_dict[cls.inc_key()] = spec_dict.pop(proxy.inc_key())
+        if proxy.def_key() in spec_dict:
+            spec_dict[cls.def_key()] = spec_dict.pop(proxy.def_key())
+        ret = proxy.build_const_args(spec_dict)
+        return ret
+
+    @classmethod
+    def _translate_kwargs(cls, kwargs):
+        """Swap mydata_type_def and mydata_type_inc for data_type_def and data_type_inc, respectively"""
+        proxy = super(BaseStorageOverride, cls)
+        kwargs[proxy.def_key()] = kwargs.pop(cls.def_key())
+        kwargs[proxy.inc_key()] = kwargs.pop(cls.inc_key())
+        return kwargs
+
+
+class CustomGroupSpec(BaseStorageOverride, GroupSpec):
+
+    @docval(*deepcopy(swap_inc_def(GroupSpec, 'CustomGroupSpec')))
+    def __init__(self, **kwargs):
+        kwargs = self._translate_kwargs(kwargs)
+        super().__init__(**kwargs)
+
+    @classmethod
+    def dataset_spec_cls(cls):
+        return CustomDatasetSpec
+
+    @docval(*deepcopy(swap_inc_def(GroupSpec, 'CustomGroupSpec')))
+    def add_group(self, **kwargs):
+        spec = CustomGroupSpec(**kwargs)
+        self.set_group(spec)
+        return spec
+
+    @docval(*deepcopy(swap_inc_def(DatasetSpec, 'CustomDatasetSpec')))
+    def add_dataset(self, **kwargs):
+        ''' Add a new specification for a subgroup to this group specification '''
+        spec = CustomDatasetSpec(**kwargs)
+        self.set_dataset(spec)
+        return spec
+
+
+class CustomDatasetSpec(BaseStorageOverride, DatasetSpec):
+
+    @docval(*deepcopy(swap_inc_def(DatasetSpec, 'CustomDatasetSpec')))
+    def __init__(self, **kwargs):
+        kwargs = self._translate_kwargs(kwargs)
+        super().__init__(**kwargs)
+
+
+class CustomSpecNamespace(SpecNamespace):
+    __types_key = 'my_data_types'
+
+    @classmethod
+    def types_key(cls):
+        return cls.__types_key